CPKAFKA-1054: Introduce batch expiry config
This is a cherry-pick from 0.11.0 to 1.0 as KIP-91 will
only land in 1.1.0 or later.

We solve the problem of expired batches not being retried
by providing the a separate config `confluent.batch.expiry.ms`
which controls how long to wait before batches are expired.
Before this patch, we would expire batches once
`request.timeout.ms` elapsed. By setting the new config
value to be very high, we ensure that batches are never expired,
which has the same effect as being retried indefinitely.