KC-827: Use TruncationOffset from the Fetch response
Simplify the fetch response by only returning a truncation offset when
necessary. In most situation one round (fetch resquest-response) of truncation
using just the offset should be enough. We expect to see multiple rounds of
truncation if there where multiple sequential leaders that wrote records to the
log but didn't get committed. We feel that this condiditon is rare.