Improve the QuorumController unit test infrastructure a bit
Rename LocalQuorumsTestEnv to QuorumControllerTestEnv and make it
independent of the LocalMetaLogManager.  This will make it easier to
test with the real Raft log code when needed.

In LocalLogManager, make it possible to limit the maximum log offset
that we will reach.  This will make it easier to test the purgatory and
the code for deferring results until records are committed.