KAFKA-18573: Add support for OAuth jwt-bearer grant type (#1598)
This change implements support for the OAuth jwt-bearer grant type.

The existing OAuth implementation supported only the OAuth client_credentials grant type, so while the changes to support the new grant type are relatively minor, a lot of refactoring was required to support both. This represents a fairly minimal approach to support the feature, but more refactoring, clean up, testing, etc. will be added after KIP-1139 is approved.