fixup: remove mapping of groups to open transactions
It was kept around in the initial PR to avoid deletion of groups with
open transactions, but it's okay to delete groups with open transactions
without any offsets, since committing the transaction is a no-op. If the
client tries to add more transactional offsets to a deleted group, we
may either recreate the group or return an error depending on the
generation in the request.