KAFKA-19431: Ensure consumer assignment consistency with subscriptions
Filter out unsubscribed topics during reconciliation.

This eliminates the window where an assignment could contain
unsubscribed topics when a member unsubscribes from a topic while it has
unrevoked partitions.

We also apply filtering in a few other cases that would arise when
client-side assignors are implemented, since new assignments would no
longer be available immediately. This is important for mixed groups,
since clients on the classic protocol will rejoin if they receive a
topic in their assignment that is no longer in their subscription.

This change does not address regex subscriptions. When regex
subscriptions are in use, we do no filtering when the regex is not
resolved, since we do not know which topics are part of the
subscription.