KAFKA-19563: improve the add controller doc (#20261)
JIRA: [KAFKA-19563](https://issues.apache.org/jira/browse/KAFKA-19563)
Improve the add-controller doc to notify users they have to add the
configs to be passed to admin client into the controller.properties
before the improvement is done.
```
--command-config COMMAND_CONFIG
      Property file containing configs to be passed to Admin Client.
For  add-controller,  the file is used to specify the controller
properties as well.
```

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>