KAFKA-17941: Fix TransactionStateManager handling of empty batch when loading transaction metadata (#20529)
When loading transaction metadata from a transaction log partition, if
the partition contains a segment ending with an empty batch,
"currOffset" update logic at will be skipped for the last batch. Since
"currOffset" is not advanced to next offset of last batch properly,
TransactionStateManager.loadTransactionMetadata method will be stuck in
the "while" loop.

This change fixes the issue by updating "currOffset" after processing
each batch, whether the batch is empty or not.

---------

Co-authored-by: Vincent Jiang <84371940+vincent81jiang@users.noreply.github.com>

Reviewers: Justine Olshan <jolshan@confluent.io>, Jun Rao <junrao@gmail.com>