KAFKA-19763: Parallel remote reads cause memory leak in broker (#20654)
Broker heap memory gets filled up and throws OOM error when remote reads are triggered for multiple partitions within a FETCH request.

Steps to reproduce:

1. Start a one node broker and configure LocalTieredStorage as remote
storage.
2. Create a topic with 5 partitions.
3. Produce message and ensure that few segments are uploaded to remote.
4. Start a consumer to read from those 5 partitions. Seek the offset to
beginning for 4 partitions and to end for 1 partition. This is to
simulate that the FETCH request read from both remote-log and local-log.
5. The broker crashes with the OOM error.
6. The DelayedRemoteFetch / RemoteLogReadResult references are being
held by the purgatory, so the broker crashes.

Reviewers: Luke Chen <showuon@gmail.com>, Satish Duggana
 <satishd@apache.org>