MINOR: Tidy up comments in MetadataVersion (#20697)
The PR just tidies up the comments in the MetadataVersion class to do
with using placeholder metadata versions for a future release as a way
to use features to enable share groups and streams groups as previews in
AK 4.1. The metadata versions are simply unstable 4.2 MVs now, which is
entirely usual at this stage of the release.

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>