KAFKA-19795: Mark the minOneMessage as false when delayedRemoteFetch is present in the first partition. (#20706)
When the FETCH request read from remote log for multiple partitions,
then all the partitions are marked with minOneMessage as true. This is
not the expected behavior:

Assume that the FETCH request is configured with fetchMaxBytes as 50 MB
and max.partition.fetch.bytes as 1 MB. And, the broker is hosting 5
partitions as leader for the topic. The FETCH request might try to read
the data from remote for all the 5 partitions. If the size of the
message in the topic is 30 MB, then we might be returning back 150 MB of
response instead of 30 MB due to minOneMessage set to true for all the
remote-read requests.

Mark the minOneMessage as false when delayedRemoteFetch is present in
the first partition.

Discussion thread:
https://github.com/apache/kafka/pull/20088#discussion_r2412551590

Reviewers: Luke Chen <showuon@gmail.com>, Satish Duggana
 <satishd@apache.org>