KAFKA-19779: Introduce source topic to subtopology map [2/N] (#20717)
For validating offset commits based on partitions, we need to be able to
efficiently find the subtopology from a given source topic (user input
topic or repartition topic).

We introduce this precomputed map in `StreamsTopology`, where it is
generated upon construction.

We can use the precomputed map in StreamsGroup.isSuscribedToTopic we can
use it to validate efficiently if we are still subscribed to a certain
topic.

Using `StreamTopology` here instead of `ConfiguredTopology` is an
improvement in itself, because configured topology is soft state and not
entirely reliable to always be there.

Reviewers: Matthias J. Sax <matthias@confluent.io>