KAFKA-19779: Add partitions to offset commit validation [1/N] (#20714)
This adds code to pass committed partitions into  ClassicGroup,
StreamsGroup and ConsumerGroup. These are not used in validation yet,
this is just preparation for later PRs.

The validation implementation in each group can return a validator,
that will be used to validate each partition in the offset commit.

We also add variations of streams groups tests for the tests in
OffsetMetadataManagerTest, which previously only tested the consumer
group case. Right now the commit logic is exactly the same, but we will
diverge in the following PRs.

Reviewers: David Jacot <djacot@confluent.io>