KAFKA-19827: Call ack commit callback at end of waiting calls (#20758)
The acknowledgement commit callback in the share consumer gets called on
the application thread at the start of the poll, commitSync and
commitAsync methods. Specifically in the peculiar case of using the
callback together with commitSync, the acknowledgement callback for the
committed records is called at the start of the next eligible call, even
though the information is already known at the end of the commitSync's
execution. The results are correct already, but the timing could be
improved in some situations.

Reviewers: Apoorv Mittal <apoorvmittal10@gmail.com>, Shivsundar R
<shr@confluent.io>