KAFKA-19734 Adding application tag to clientstate metric (#20766)
## Summary

As a follow-on to the improvements introduced in

[KIP-1091](https://cwiki.apache.org/confluence/display/KAFKA/KIP-1091%3A+Improved+Kafka+Streams+operator+metrics)
it would be useful to add the application-id as a tag to the
`client-state` metric. This allows Kafka Streams developers and
operators to connect metrics containing a `thread-id` (which embeds the
`application-id`) across separate deployments of Kafka Streams
instances, which are members of the same logical application.

Reviewers: Bill Bejeck<bbejeck@apache.org> Matthias
 Sax<mjsax@apache.org>