KAFKA-19449: Don't consider partitions unreleased when pending revocation by the current member
During reconciliation, we transition members to UNRELEASED_PARTITIONS
when the target assignment contains a partition that is owned by another
member. Partitions are considered owned by another member if they are in
a member's assignment or partitions pending revocation. Ownership is
only updated after reconciliation has finished.

Fix a bug where we consider a partition owned by another member when it
is actually in the current member's list of partitions pending
revocation. In some cases, this skips a transition to
UNRELEASED_PARTITIONS and allows the current assignment to converge to
the target assignment one heartbeat sooner.