MINOR: Fix time comparison with appendLingerMs in CoordinatorRuntime#maybeFlushCurrentBatch (#20773)
This PR fixed the time comparison logic in
CoordinatorRuntime#maybeFlushCurrentBatch to ensure that the batch is
flushed when the elapsed time since appendTimeMs exceeds the
appendLingerMs parameter.

This issue is also mentioned
[here](https://github.com/apache/kafka/pull/20653/files#r2442452104).

The fix for this issue was originally in [this
PR](https://github.com/apache/kafka/pull/20739) in the trunk branch,
which was backported to the 4.0 branch.

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>