KAFKA-16024: SaslPlaintextConsumerTest#testCoordinatorFailover is flaky (#20774)
- `broker.session.timeout.ms` defaults to 9s. When a broker goes
offline, the group coordinator may take up to this long to be
re-elected.
- The commit callback retry timeout is currently 10 seconds, which
leaves very little buffer. If the metadata hasn’t refreshed yet, the
consumer may still send an OFFSET_COMMIT request to the offline
coordinator, leading to transient failures.

This patch enable `controlled.shutdown.enable` to allow the broker to
notify the  controller before shutting down.  This speeds up the test by
triggering  an immediate failover instead of waiting for the  broker
session timeout  (default: 9s) to expire.

Reviewers: TaiJuWu <tjwu1217@gmail.com>, PoAn Yang <payang@apache.org>