KAFKA-19813 Incorrect jitter value in StreamsGroupHeartbeatRequestManager and AbstractHeartbeatRequestManager (#20748)
In StreamsGroupHeartbeatRequestManager and
AbstractHeartbeatRequestManager  we use max.poll.interval.ms (default
300000) as the jitter value. Although RequestState.remainingBackoffMs
guards with Math.max(0, …), which won’t make the value become negative,
using a jitter that isn’t in (0–1) is unexpected.

In addition, we should validate that ExponentialBackoff receives a
jitter strictly within (0, 1) to prevent this scenario in the future.

Reviewers: Chia-Ping Tsai <chia7712@gmail.com>