KAFKA-19426 TopicBasedRemoteLogMetadataManager's initial should happen after the broker ready (KIP-1197) (#20203)
refer to https://github.com/apache/kafka/pull/20008. I submit another
proposal to solve the issue which can also help to solve the following
issue:   the
DEFAULT_REMOTE_LOG_METADATA_INITIALIZATION_RETRY_MAX_TIMEOUT_MS is hard
to be defined due to that it relay on the startup process.

If it the value is not well defined. the initial will fail and cause
local disk never get deleted.  （propose
https://github.com/apache/kafka/pull/20007 to prevent it.）

Different env's kafka process need different value. It is caused by
followed reason:  When restarting broker. The connection to query/create
topic in "TopicBasedRemoteLogMetadataManager#initializeResources"will
fail until the broker's self get ready. The detail reason can refer to
:https://github.com/apache/kafka/pull/20007#issuecomment-3027094794

So propose this change to improve it. After the improve it and default
value is easy to be defined. I don't need to set a very big value to all
our kafkas.

Reviewers: Yung <yungyung7654321@gmail.com>, Kamal Chandraprakash
 <kamal.chandraprakash@gmail.com>, Chia-Ping Tsai <chia7712@gmail.com>