KAFKA-19846: Remove uuid to string conversion and vice-versa for memberId on broker (#20819)
### About
This PR aims to resolve the bug when we have '/' or '+' in `memberId`
within share fetch and share acknowledge RPCs. The bug arises primarily
because we do a `string -> Uuid` conversion and vice-versa in the broker
for `memberId`. When a string containing '/' or '+' is converted to
Uuid, the `Base64.getUrlDecoder()` is unable to decode the string to a
Uuid and gives an exception which is the reason for the bug.
On analysis, it looks like there is no efficiency added due to this
inter-conversion. Hence, we have gotten rid of the conversion from
`string -> Uuid` in `KafkaApis` for `memberId`. Thus, everywhere on the
broker memberId always treated as string not a Uuid

While fixing this, I noticed that we don't need mapping in
`ShareSessionContext` class, hence removed it.

### Testing
Added a few unit tests to validate the new functionality

Reviewers: Apoorv Mittal <apoorvmittal10@gmail.com>, Andrew Schofield <aschofield@confluent.io>