KAFKA-19446: Add Transaction Version to request builder [2/3] (#20868)
In this patch for KIP-1228, we add a new WriteTxnMarkersRequest builder
constructor  that includes the    transaction versions for each marker.
The  transaction version is already stored in   the metadata on the
transaction coordinator. We now extract this version   information and
pass it to the builder via the TxnMarkerEntry, which propagates it to
partition leaders.

Since the TransactionVersion field is marked as ignorable, we always
include it when   building the request. The field will be automatically
omitted without any errors, during serialization   if the request
version is < 2, ensuring backward compatibility with brokers that   only
support version 1.  

Reviewers: Justine Olshan <jolshan@confluent.io>