KAFKA-19829: Implement group-level initial rebalance delay (#20755)
During testing, an artifact of the new rebalance protocol showed up. In
some cases, the first joining member gets all active tasks assigned, and
is slow to revoke the tasks after more member has joined the group. This
affects in particular cases where the first member is slow (possibly
overloaded in the case of cloudlimits benchmarks) and there are a lot of
tasks to be assigned.

To help with this situation, we want to introduce a new group-specific
configuration to delay the initial rebalance.

Main changes:
- add 3s as default delay to `GroupConfig` & `GroupCoordinatorConfig`
- add `is_scheduled method` to timer
- add rebalance delay logic
  - check if it's initial rebalance by using `bumpGroup` & `groupEpoch
== 0`
  - if it's initial rebalance then trigger the timer and bump the group
by two
  - when the timer is end, compute a new target assignment
- add related tests
- ignore epoch check in `DescribeStreamsGroupTest` since it's flaky now
- add `setGroupStreamsInitialRebalanceDelay` to EmbeddedKafkaCluster` so
we can manully set the config.
- Set delay to 0 in most of the tests

Reviewers: Lucas Brutschy <lbrutschy@confluent.io>, Sean Quah
 <squah@confluent.io>