KAFKA-19781: Fix to not update positions for partitions being revoked (#20914)
Fix to avoid initializing positions for partitions being revoked, as it
is unneeded (we do not allow fetching from partitions being revoked),
and could lead to NoOffsetForPartitionException on a partition that is
already being revoked (this is confusing for applications).

This is the behaviour we already had for fetch, just applying it to
update positions to align.

This gap was surfaced on edge cases of partitions being assigned and
revoked right away.

Reviewers: Andrew Schofield <aschofield@confluent.io>, Lucas Brutschy
 <lbrutschy@confluent.io>