MINOR: Revert create local state Standbys on start (#16922)" (#20981)
This reverts commit 571f5081
(https://github.com/apache/kafka/pull/16922) of an incomplete feature.

PR https://github.com/apache/kafka/pull/16922 is part of

[KIP-1035](https://cwiki.apache.org/confluence/display/KAFKA/KIP-1035%3A+StateStore+managed+changelog+offsets).
In particular,  on starting a Kafka Streams instance, if it has
pre-existing state, the  state stores are initialized on the main
thread. Part of this  initialization registers the stateful metrics with
the JMX thread-id tag  of main. This breaks the KIP-1076 implementation
where need to  register metrics with thread-id tags of
xxxStreamThread-N. This is  necessary due to the fact that the
StreamsMetric is a singleton shared  by all StreamThread instances, so
we need to make sure only add  metrics for the current StreamThread
otherwise duplicate metrics are  registered. This PR reverts the changes
until a fix is implemented,  allowing the individual StreamThreads to
register the metrics.

Reviewers: Matthias Sax<mjsax@apache.org>