KAFKA-19589: Check for ability to skip position validation in application thread when collecting buffered data (#20324)
Introduces `PositionsValidator` which queries the assignment data  from
the `SubscriptionState` from the  application thread, allowing
`AsyncKafkaConsumer` to potentially avoid the need to wait.

The impetus for this change is the observation of these two points:

1. Waiting for the completion of
`OffsetsRequestManager.updateFetchPositions()` on the application thread
means either busy-waiting or blocking, either of which add significant
(~60%) CPU load to the `AsyncKafkaConsumer` compared to the
`ClassicKafkaConsumer`
2. In testing, data shows that 99.99+% of the time that
`OffsetsRequestManager.updateFetchPositions()` is called, the partitions
are up-to-date and there is no need to fetch offsets.

This patch allows the check for stable partitions to be made in the
application thread, resulting in far less waiting in the critical path
of `AsyncKafkaConsumer.poll()`.

Reviewers: Lianet Magrans <lmagrans@confluent.io>, Andrew Schofield
 <aschofield@confluent.io>