MINOR: Refactor OffsetFetch path (#21009)
The `GroupCoordinator` interface has two methods to fetch offsets:
`fetchOffsets` and `fetchAllOffsets`. They have the exact same signature
and the implementation in `GroupCoordinatorService` is exactly the same,
modulo the name of the operation. The path refactors the path to
simplify it and reuse more code. We could further refactor `KafkaApis`
but let's do this in a follow-up in order to keep this change small and
simple.

Reviewers: Sean Quah <squah@confluent.io>, Lianet Magrans
<lmagrans@confluent.io>